feat-comments and class validating
Added a comment section to the Post Page, and started to work with class validations.

+++ fullstack-blog/server/src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Post } from './entities/post.entity';
@@ -11,46 +11,73 @@ export class PostsService {
   constructor(
     @InjectRepository(Post)
     private postsRepository: Repository<Post>,
-  ){}
+  ) { }
 
   async create(createPostDto: CreatePostDto, author_id: number): Promise<Post> {
-    const newPost = this.postsRepository.create({...createPostDto, author_id});
+    const newPost = this.postsRepository.create({ ...createPostDto, author_id });
     return this.postsRepository.save(newPost);
   }
 
   async findAll(): Promise<Post[]> {
     const posts = await this.postsRepository.find();
-    if (posts.length === 0 ){
-      throw new NotFoundException(`Not a single post was found. Please register/create one. :)`);
-    }
+    
     return posts;
   }
 
   async findOne(id: number): Promise<Post> {
-    const post = await this.postsRepository.findOneBy({id});
-    if (!post){
+    const post = await this.postsRepository.findOne({
+      where: { id },
+      relations: ['user'],
+      select: {
+        id: true,
+        title: true,
+        content: true,
+        author_id: true,
+        created_at: true,
+        updated_at: true,
+        // Select only some columns so it doesn't return things like the hashed password.
+        user: {
+          id: true,
+          name: true,
+          sur_name: true,
+          email: true,
+        },
+      },
+    },);
+    if (!post) {
       throw new NotFoundException(`The post with id ${id} isn't currently registred, please check if this is the right id. :)`);
     }
     return post;
   }
 
-  async update(id: number, updatePostDto: UpdatePostDto): Promise<Post> {
+  async update(id: number, updatePostDto: UpdatePostDto, user_id: number): Promise<Post> {
     const post = await this.postsRepository.preload({
       id: id,
       ...updatePostDto,
     })
-    if (!post){
+    if (!post) {
       throw new NotFoundException(`The post with id ${id} couldn't be found for updating it's data, please check if it is registred. :)`);
     }
-    
+
+    if (user_id !== post.author_id) {
+      throw new ForbiddenException('Only who created the post can alter it! :|');
+    }
+
+
     return await this.postsRepository.save(post);
   }
 
-  async remove(id: number): Promise<string> {
-    const deleted = await this.postsRepository.delete(id);
-    if (deleted.affected === 0){
-      throw new NotFoundException(`The post ${id} was already deleted or doesn't exists. Either way, it's gone. :)`);
+  async remove(id: number, user_id: number): Promise<string> {
+    const post = await this.postsRepository.findOneBy({ id });
+    if (!post) {
+      throw new NotFoundException(`The post with id ${id} couldn't be found for deleting it's data, please check if it is registred. :)`);
     }
+
+    if (user_id !== post.author_id) {
+      throw new ForbiddenException('Only who created the post can remove it! :|');
+    }
+
+    await this.postsRepository.delete(id);
     return `The post ${id} was succefully deleted. :)`;
   }
 }
